third_party/blink/web_tests/fast/scrolling/scrollbars/mouse-scrolling-on-div-scrollbar-rtl.html

<!DOCTYPE html>
<title>Tests mouse interactions on a non-custom composited div scrollbar.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../resources/blink-coordinates-util.js"></script>
<script src="../../../resources/scrollbar-util.js"></script>
<script src="resources/mouse-scrolling-on-div-scrollbar.js"></script>
<style>
  .appearance {
    width: 100px;
    height: 100px;
    overflow: scroll;
    border: 1px solid black;
  }

  .standardLocation {
    position: absolute;
    top: 100px;
    left: 100px;
  }

  .space {
    height: 200px;
    width: 200px;
  }
</style>

<!-- Composited non-custom scroller -->
<div id="standard" class="appearance standardLocation" style="direction: rtl;">
  <div class="space"></div>
</div>

<script>
  if (window.internals)
    internals.settings.setScrollAnimatorEnabled(false);

  const scroller = document.getElementById("standard");
  const TRACK_WIDTH = calculateScrollbarThickness();

  assert_equals(scroller.clientHeight, scroller.clientWidth,
    "This test assumes that the height and width of 'scroller' are equivalent. If this changes please update SCROLL_AMOUNT to be X/Y specific");

  const params = {
    scroller: scroller,
    TRACK_WIDTH: TRACK_WIDTH,
    BUTTON_WIDTH: TRACK_WIDTH,
    SCROLL_CORNER: TRACK_WIDTH,
    SCROLL_AMOUNT: getScrollbarButtonScrollDelta(scroller).y
  }

  promise_test(async () => {
    await testArrows(params);
  }, "Test mouse click on non-custom composited div scrollbar arrows (RTL).");
  promise_test(async () => {
    await testTrackparts(params);
  }, "Test mouse click on non-custom composited div scrollbar empty trackparts (RTL).");
</script>
